+++ test/Veggerby.Boards.Tests/Backgammon/BackgammonGameBuilderTests.cs
@@ -1,9 +1,8 @@
 using System;
 
-
-using Veggerby.Boards.Backgammon;
 using Veggerby.Boards;
 using Veggerby.Boards.Artifacts;
+using Veggerby.Boards.Backgammon;
 using Veggerby.Boards.Tests.Utils;
 
 namespace Veggerby.Boards.Tests.Backgammon;